server/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
PyJWT==2.10.1
sqlparse==0.5.5
django-redis==6.0.0
pyotp==2.9.0
qrcode==8.2
redis==7.1.1
requests==2.32.5
django-ratelimit==4.1.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0
websockets==13.1
